fix: rename preview-api-keys to preview-api-key
