+++ lib/adapters/REST/endpoints/index.ts
@@ -13,7 +13,7 @@ import * as Organization from './organization'
 import * as OrganizationInvitation from './organization-invitation'
 import * as OrganizationMembership from './organization-membership'
 import * as PersonalAccessToken from './personal-access-token'
-import * as PreviewApiKey from './preview-api-keys'
+import * as PreviewApiKey from './preview-api-key'
 import * as Role from './role'
 import * as ScheduledAction from './scheduled-action'
 import * as Snapshot from './snapshot'
